Move linking to utils.jira and use ModuleImporter
